+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 59.0 - 2023-07-14
+### Added
+- Map data to the Dawn of the Infinite dungeon and Time Rift event
+
+### Changed
+- Moved the Memory of Scholomance and Leaders of Scholomance achievements to Classic - Dungeons - Scholomance
+- Moved Wards of the Dread Citadel, 'Never Bothered, Anyway' and Putting Wilhelm Out of Business to Classic - Raids - Naxxramas
+- Moderately reduced the memory and Exported Map file sizes
+
+### Fixed
+- Removed rounding on Event Reminder end times; end times should now better match the actual timers
+- Community Feast and Time Rifts should no longer bug out when in the last minute of the count down
+
 ## 58.1 - 2023-07-12
 ### Fixed
 - Tracking achievements should once again work